requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-MySQL==1.6.0
fonttools==4.55.8
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mysql-connector-python==9.2.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytrends==4.9.2
pytz==2025.1
requests==2.32.3
seaborn==0.13.2
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
